refactor: Moves the select-logic into a new helper file so it can be more easily used for checking synced-values as well
